pull out PropagateEquals tests
