+++ x-pack/plugin/esql/src/test/java/org/elasticsearch/xpack/esql/optimizer/OptimizerRulesTests.java
@@ -88,295 +88,6 @@ public class OptimizerRulesTests extends ESTestCase {
 
     // Test Propagate Equals
 
-    // a == 1 AND a == 2 -> FALSE
-    public void testDualEqualsConjunction() {
-        FieldAttribute fa = getFieldAttribute();
-        Equals eq1 = equalsOf(fa, ONE);
-        Equals eq2 = equalsOf(fa, TWO);
-
-        PropagateEquals rule = new PropagateEquals();
-        Expression exp = rule.rule(new And(EMPTY, eq1, eq2));
-        assertEquals(FALSE, exp);
-    }
-
-    // 1 < a < 10 AND a == 10 -> FALSE
-    public void testEliminateRangeByEqualsOutsideInterval() {
-        FieldAttribute fa = getFieldAttribute();
-        Equals eq1 = equalsOf(fa, new Literal(EMPTY, 10, DataType.INTEGER));
-        Range r = rangeOf(fa, ONE, false, new Literal(EMPTY, 10, DataType.INTEGER), false);
-
-        PropagateEquals rule = new PropagateEquals();
-        Expression exp = rule.rule(new And(EMPTY, eq1, r));
-        assertEquals(FALSE, exp);
-    }
-
-    // a != 3 AND a = 3 -> FALSE
-    public void testPropagateEquals_VarNeq3AndVarEq3() {
-        FieldAttribute fa = getFieldAttribute();
-        NotEquals neq = notEqualsOf(fa, THREE);
-        Equals eq = equalsOf(fa, THREE);
-
-        PropagateEquals rule = new PropagateEquals();
-        Expression exp = rule.rule(new And(EMPTY, neq, eq));
-        assertEquals(FALSE, exp);
-    }
-
-    // a != 4 AND a = 3 -> a = 3
-    public void testPropagateEquals_VarNeq4AndVarEq3() {
-        FieldAttribute fa = getFieldAttribute();
-        NotEquals neq = notEqualsOf(fa, FOUR);
-        Equals eq = equalsOf(fa, THREE);
-
-        PropagateEquals rule = new PropagateEquals();
-        Expression exp = rule.rule(new And(EMPTY, neq, eq));
-        assertEquals(Equals.class, exp.getClass());
-        assertEquals(eq, exp);
-    }
-
-    // a = 2 AND a < 2 -> FALSE
-    public void testPropagateEquals_VarEq2AndVarLt2() {
-        FieldAttribute fa = getFieldAttribute();
-        Equals eq = equalsOf(fa, TWO);
-        LessThan lt = lessThanOf(fa, TWO);
-
-        PropagateEquals rule = new PropagateEquals();
-        Expression exp = rule.rule(new And(EMPTY, eq, lt));
-        assertEquals(FALSE, exp);
-    }
-
-    // a = 2 AND a <= 2 -> a = 2
-    public void testPropagateEquals_VarEq2AndVarLte2() {
-        FieldAttribute fa = getFieldAttribute();
-        Equals eq = equalsOf(fa, TWO);
-        LessThanOrEqual lt = lessThanOrEqualOf(fa, TWO);
-
-        PropagateEquals rule = new PropagateEquals();
-        Expression exp = rule.rule(new And(EMPTY, eq, lt));
-        assertEquals(eq, exp);
-    }
-
-    // a = 2 AND a <= 1 -> FALSE
-    public void testPropagateEquals_VarEq2AndVarLte1() {
-        FieldAttribute fa = getFieldAttribute();
-        Equals eq = equalsOf(fa, TWO);
-        LessThanOrEqual lt = lessThanOrEqualOf(fa, ONE);
-
-        PropagateEquals rule = new PropagateEquals();
-        Expression exp = rule.rule(new And(EMPTY, eq, lt));
-        assertEquals(FALSE, exp);
-    }
-
-    // a = 2 AND a > 2 -> FALSE
-    public void testPropagateEquals_VarEq2AndVarGt2() {
-        FieldAttribute fa = getFieldAttribute();
-        Equals eq = equalsOf(fa, TWO);
-        GreaterThan gt = greaterThanOf(fa, TWO);
-
-        PropagateEquals rule = new PropagateEquals();
-        Expression exp = rule.rule(new And(EMPTY, eq, gt));
-        assertEquals(FALSE, exp);
-    }
-
-    // a = 2 AND a >= 2 -> a = 2
-    public void testPropagateEquals_VarEq2AndVarGte2() {
-        FieldAttribute fa = getFieldAttribute();
-        Equals eq = equalsOf(fa, TWO);
-        GreaterThanOrEqual gte = greaterThanOrEqualOf(fa, TWO);
-
-        PropagateEquals rule = new PropagateEquals();
-        Expression exp = rule.rule(new And(EMPTY, eq, gte));
-        assertEquals(eq, exp);
-    }
-
-    // a = 2 AND a > 3 -> FALSE
-    public void testPropagateEquals_VarEq2AndVarLt3() {
-        FieldAttribute fa = getFieldAttribute();
-        Equals eq = equalsOf(fa, TWO);
-        GreaterThan gt = greaterThanOf(fa, THREE);
-
-        PropagateEquals rule = new PropagateEquals();
-        Expression exp = rule.rule(new And(EMPTY, eq, gt));
-        assertEquals(FALSE, exp);
-    }
-
-    // a = 2 AND a < 3 AND a > 1 AND a != 4 -> a = 2
-    public void testPropagateEquals_VarEq2AndVarLt3AndVarGt1AndVarNeq4() {
-        FieldAttribute fa = getFieldAttribute();
-        Equals eq = equalsOf(fa, TWO);
-        LessThan lt = lessThanOf(fa, THREE);
-        GreaterThan gt = greaterThanOf(fa, ONE);
-        NotEquals neq = notEqualsOf(fa, FOUR);
-
-        PropagateEquals rule = new PropagateEquals();
-        Expression and = Predicates.combineAnd(asList(eq, lt, gt, neq));
-        Expression exp = rule.rule((And) and);
-        assertEquals(eq, exp);
-    }
-
-    // a = 2 AND 1 < a < 3 AND a > 0 AND a != 4 -> a = 2
-    public void testPropagateEquals_VarEq2AndVarRangeGt1Lt3AndVarGt0AndVarNeq4() {
-        FieldAttribute fa = getFieldAttribute();
-        Equals eq = equalsOf(fa, TWO);
-        Range range = rangeOf(fa, ONE, false, THREE, false);
-        GreaterThan gt = greaterThanOf(fa, new Literal(EMPTY, 0, DataType.INTEGER));
-        NotEquals neq = notEqualsOf(fa, FOUR);
-
-        PropagateEquals rule = new PropagateEquals();
-        Expression and = Predicates.combineAnd(asList(eq, range, gt, neq));
-        Expression exp = rule.rule((And) and);
-        assertEquals(eq, exp);
-    }
-
-    // a = 2 OR a > 1 -> a > 1
-    public void testPropagateEquals_VarEq2OrVarGt1() {
-        FieldAttribute fa = getFieldAttribute();
-        Equals eq = equalsOf(fa, TWO);
-        GreaterThan gt = greaterThanOf(fa, ONE);
-
-        PropagateEquals rule = new PropagateEquals();
-        Expression exp = rule.rule(new Or(EMPTY, eq, gt));
-        assertEquals(gt, exp);
-    }
-
-    // a = 2 OR a > 2 -> a >= 2
-    public void testPropagateEquals_VarEq2OrVarGte2() {
-        FieldAttribute fa = getFieldAttribute();
-        Equals eq = equalsOf(fa, TWO);
-        GreaterThan gt = greaterThanOf(fa, TWO);
-
-        PropagateEquals rule = new PropagateEquals();
-        Expression exp = rule.rule(new Or(EMPTY, eq, gt));
-        assertEquals(GreaterThanOrEqual.class, exp.getClass());
-        GreaterThanOrEqual gte = (GreaterThanOrEqual) exp;
-        assertEquals(TWO, gte.right());
-    }
-
-    // a = 2 OR a < 3 -> a < 3
-    public void testPropagateEquals_VarEq2OrVarLt3() {
-        FieldAttribute fa = getFieldAttribute();
-        Equals eq = equalsOf(fa, TWO);
-        LessThan lt = lessThanOf(fa, THREE);
-
-        PropagateEquals rule = new PropagateEquals();
-        Expression exp = rule.rule(new Or(EMPTY, eq, lt));
-        assertEquals(lt, exp);
-    }
-
-    // a = 3 OR a < 3 -> a <= 3
-    public void testPropagateEquals_VarEq3OrVarLt3() {
-        FieldAttribute fa = getFieldAttribute();
-        Equals eq = equalsOf(fa, THREE);
-        LessThan lt = lessThanOf(fa, THREE);
-
-        PropagateEquals rule = new PropagateEquals();
-        Expression exp = rule.rule(new Or(EMPTY, eq, lt));
-        assertEquals(LessThanOrEqual.class, exp.getClass());
-        LessThanOrEqual lte = (LessThanOrEqual) exp;
-        assertEquals(THREE, lte.right());
-    }
-
-    // a = 2 OR 1 < a < 3 -> 1 < a < 3
-    public void testPropagateEquals_VarEq2OrVarRangeGt1Lt3() {
-        FieldAttribute fa = getFieldAttribute();
-        Equals eq = equalsOf(fa, TWO);
-        Range range = rangeOf(fa, ONE, false, THREE, false);
-
-        PropagateEquals rule = new PropagateEquals();
-        Expression exp = rule.rule(new Or(EMPTY, eq, range));
-        assertEquals(range, exp);
-    }
-
-    // a = 2 OR 2 < a < 3 -> 2 <= a < 3
-    public void testPropagateEquals_VarEq2OrVarRangeGt2Lt3() {
-        FieldAttribute fa = getFieldAttribute();
-        Equals eq = equalsOf(fa, TWO);
-        Range range = rangeOf(fa, TWO, false, THREE, false);
-
-        PropagateEquals rule = new PropagateEquals();
-        Expression exp = rule.rule(new Or(EMPTY, eq, range));
-        assertEquals(Range.class, exp.getClass());
-        Range r = (Range) exp;
-        assertEquals(TWO, r.lower());
-        assertTrue(r.includeLower());
-        assertEquals(THREE, r.upper());
-        assertFalse(r.includeUpper());
-    }
-
-    // a = 3 OR 2 < a < 3 -> 2 < a <= 3
-    public void testPropagateEquals_VarEq3OrVarRangeGt2Lt3() {
-        FieldAttribute fa = getFieldAttribute();
-        Equals eq = equalsOf(fa, THREE);
-        Range range = rangeOf(fa, TWO, false, THREE, false);
-
-        PropagateEquals rule = new PropagateEquals();
-        Expression exp = rule.rule(new Or(EMPTY, eq, range));
-        assertEquals(Range.class, exp.getClass());
-        Range r = (Range) exp;
-        assertEquals(TWO, r.lower());
-        assertFalse(r.includeLower());
-        assertEquals(THREE, r.upper());
-        assertTrue(r.includeUpper());
-    }
-
-    // a = 2 OR a != 2 -> TRUE
-    public void testPropagateEquals_VarEq2OrVarNeq2() {
-        FieldAttribute fa = getFieldAttribute();
-        Equals eq = equalsOf(fa, TWO);
-        NotEquals neq = notEqualsOf(fa, TWO);
-
-        PropagateEquals rule = new PropagateEquals();
-        Expression exp = rule.rule(new Or(EMPTY, eq, neq));
-        assertEquals(TRUE, exp);
-    }
-
-    // a = 2 OR a != 5 -> a != 5
-    public void testPropagateEquals_VarEq2OrVarNeq5() {
-        FieldAttribute fa = getFieldAttribute();
-        Equals eq = equalsOf(fa, TWO);
-        NotEquals neq = notEqualsOf(fa, FIVE);
-
-        PropagateEquals rule = new PropagateEquals();
-        Expression exp = rule.rule(new Or(EMPTY, eq, neq));
-        assertEquals(NotEquals.class, exp.getClass());
-        NotEquals ne = (NotEquals) exp;
-        assertEquals(FIVE, ne.right());
-    }
-
-    // a = 2 OR 3 < a < 4 OR a > 2 OR a!= 2 -> TRUE
-    public void testPropagateEquals_VarEq2OrVarRangeGt3Lt4OrVarGt2OrVarNe2() {
-        FieldAttribute fa = getFieldAttribute();
-        Equals eq = equalsOf(fa, TWO);
-        Range range = rangeOf(fa, THREE, false, FOUR, false);
-        GreaterThan gt = greaterThanOf(fa, TWO);
-        NotEquals neq = notEqualsOf(fa, TWO);
-
-        PropagateEquals rule = new PropagateEquals();
-        Expression exp = rule.rule((Or) Predicates.combineOr(asList(eq, range, neq, gt)));
-        assertEquals(TRUE, exp);
-    }
-
-    // a == 1 AND a == 2 -> nop for date/time fields
-    public void testPropagateEquals_ignoreDateTimeFields() {
-        FieldAttribute fa = getFieldAttribute("a", DataType.DATETIME);
-        Equals eq1 = equalsOf(fa, ONE);
-        Equals eq2 = equalsOf(fa, TWO);
-        And and = new And(EMPTY, eq1, eq2);
-
-        PropagateEquals rule = new PropagateEquals();
-        Expression exp = rule.rule(and);
-        assertEquals(and, exp);
-    }
-
-    // 1 <= a < 10 AND a == 1 -> a == 1
-    public void testEliminateRangeByEqualsInInterval() {
-        FieldAttribute fa = getFieldAttribute();
-        Equals eq1 = equalsOf(fa, ONE);
-        Range r = rangeOf(fa, ONE, true, new Literal(EMPTY, 10, DataType.INTEGER), false);
-
-        PropagateEquals rule = new PropagateEquals();
-        Expression exp = rule.rule(new And(EMPTY, eq1, r));
-        assertEquals(eq1, exp);
-    }
     //
     // Null folding
